Dockerise the mongoDB
Dockerise the mongoDB into containers and intergrate with KT application
