+++ .dockerignore
@@ -1,0 +1,3 @@
+...
+.gitignore
+.env
