--- Booking/public/css/main.css
@@ -1,137 +1,0 @@
-body {
-    background: #dddddd;
-    margin: 24px;
-  }
-  
-  .timer {
-    font-family: sans-serif;
-    display: inline-block;
-    padding: 24px 32px;
-    border-radius: 30px;
-    background: white;
-  }
-  
-  .timer__part {
-    font-size: 36px;
-    font-weight: bold;
-  }
-  
-  .timer__btn {
-    width: 50px;
-    height: 50px;
-    margin-left: 16px;
-    border-radius: 50%;
-    border: none;
-    color: white;
-    background: #8208e6;
-    cursor: pointer;
-  }
-  
-  .timer__btn--start {
-    background: #00b84c;
-  }
-  
-  .timer__btn--stop {
-    background: #ff0256;
-  }
-  
-  .title {
-    font-family: sans-serif;
-    display: inline-block;
-    padding: 24px 32px;
-    border-radius: 30px;
-    background: white;
-  }
-  
-  .title__part {
-    font-size: 36px;
-    font-weight: bold;
-  }
-
-  .body {
-    align-items: center;
-    justify-content: center;
-  }
-
-  .seat {
-    background-color: white;
-    height: 12px;
-    width: 15px;
-    margin: 3px;
-    border-top-left-radius: 10px;
-    border-top-right-radius: 10px;
-  }
-
-  .seat.selected {
-    background-color: aqua;
-  }
-
-  .seat.occupied {
-    background-color:dimgrey ;
-  }
-
-  .seat:not(.occupied):hover {
-    cursor: pointer;
-    transform: scale(1.2);
-  }
-
-  .seat:nth-of-type(2) {
-    margin-right: 18px;
-  }
-
-  .seat:nth-last-of-type(2) {
-    margin-left: 18px;
-  }
-
-  .showcase .seat:not(.seat.occupied):hover {
-    cursor: default;
-    transform: scale(1);
-  }
-
-  .showcase {
-    background-color: rgba(0,0,0,0.1);
-    padding: 5px 10px;
-    border-radius: 5px;
-    color: #777;
-    list-style-type: none;
-    display: flex;
-    justify-content:  space-evenly;
-  }
-
-  .showcase li {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-
-  .showcase li small {
-    margin-left: 20px;
-  }
-
-  .row {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-
-  .screen {
-    background-color: white;
-    height: 70px;
-    width: 100%;
-    margin: 15px 0;
-    transform: rotateX(-45deg);
-    box-shadow: 0 3px 10px rgba(255, 255, 255, 0.75);
-  }
-
-  .container {
-    perspective: 1000px;
-    margin-bottom: 30px;
-  }
-
-  p.text {
-    margin: 5px 0;
-  }
-
-  p.text span {
-    color: aqua
-  }
