--- Booking/public/index.html
@@ -1,123 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Booking</title>
-    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-    <link rel="stylesheet" href="css/main.css">
-    <!--<script type="module" src="backend/index.js"></script>-->
-</head>
-<body>
-    <div class="timer"></div>
-    <div class="title">
-        <span id="theater">THEATER</span> / <span id="movie">MOVIE</span> 
-    </div>
-
-    <ul class="showcase">
-        <li>
-            <div class="seat">
-                <small>N/A</small>
-            </div>
-        </li>
-
-        <li>
-            <div class="seat selected">
-                <small>Selected</small>
-            </div>
-        </li>
-
-        <li>
-            <div class="seat occupied">
-                <small>Occupied</small>
-            </div>
-        </li>
-    </ul>
-
-    <div class="container">
-        <div class="screen"></div>
-        <div class="row">
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-        </div>
-
-        <div class="row">
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-        </div>
-
-        <div class="row">
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-        </div>
-
-        <div class="row">
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-        </div>
-
-        <div class="row">
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-            <div class="seat"></div>
-        </div>
-    </div>
-
-    <p class="text">
-        You have selected <span id="count">0</span> seats.
-    </p>
-
-    <button type="button" class="booking__btn booking__btn--confirm">
-        Confirm
-    </button>
-    <button type="button" class="booking__btn booking__btn--cancel">
-        Cancel
-    </button>
-    <button type="button" class="test__btn test__btn--test">
-        Test
-    </button>
-
-    <script type="module" src="js/main.js"></script>
-    <script src="js/seat.js"></script>
-</body>
-</html>
