--- Booking/public/js/main.js
@@ -1,10 +1,0 @@
-import Timer from "./Timer.js"
-import Title from "./Title.js"
-
-new Timer(
-    document.querySelector(".timer")
-);
-
-new Title(
-    document.querySelector(".title")
-);
