+++ docker-compose.yml
@@ -1,0 +1,61 @@
+version: "3"
+
+services:
+  booking-svc:
+    image: kt/booking-svc
+    container_name: booking-svc
+    restart: unless-stopped
+    env_file: .env
+    environment:
+      - MONGO_USERNAME=rootuser
+      - MONGO_PASSWORD=rootpass
+      - MONGO_HOSTNAME=mongodb
+      - MONGO_PORT=27017
+      - MONGO_DB=mongodb
+    expose:
+      - 8989
+    ports:
+      - "8989:3000"
+    volumes:
+      - .:/home/appuser/app
+      - node_modules:/home/appuser/app/node_modules
+    networks:
+      - app-network
+    depends_on:
+      - mongodb
+
+  mongodb:
+    image: mongo
+    container_name: mongodb
+    ports:
+      - 27017:27017
+    environment:
+      - MONGO_INITDB_ROOT_USERNAME=rootuser
+      - MONGO_INITDB_ROOT_PASSWORD=rootpass
+    volumes:
+      - dbdata:/data/db
+    networks:
+      - app-network
+
+  mongo-express:
+    image: mongo-express
+    container_name: mongo-express
+    ports:
+      - 8081:8081
+    environment:
+      - ME_CONFIG_MONGODB_ADMINUSERNAME=rootuser
+      - ME_CONFIG_MONGODB_ADMINPASSWORD=rootpass
+      - ME_CONFIG_MONGODB_SERVER=mongodb
+    depends_on:
+      - mongodb
+    networks:
+      - app-network
+
+networks:
+  app-network:
+    driver: bridge
+
+
+volumes:
+  dbdata:
+  node_modules:
